8287593: ShortResponseBody could be made more resilient to rogue connections
Since ReplyingServer binds to an open port on the system where tests
run, it is possible some other program probes that port while the test
is running. If the probe request does not have request params, the
request is invalid and a test assertion fails. This commit replaces that
assertion with returning a 400 response.