Undoing commit
If we try adding this to Throwable, it gets a little ugly because you can't have an unreported exception. By reporting the Exception though, you now force every method that calls this method to add a throws clause. This poisons things downstream and inhibits the main goal which is to concisely and precisely validate precondition. We have exchanged precise for concise, and therefore, no good.

So, I am undoing this commit.