8290730: compiler/vectorization/TestAutoVecIntMinMax.java failed with "IRViolationException: There were one or multiple IR rule failures."
The IR test - TestAutoVecIntMinMax.java was introduced in https://bugs.openjdk.org/browse/JDK-8288107 to test IR generation of MaxV and MinV nodes when the MinI/MaxI nodes are auto-vectorized.
However, the corresponding vector ISA support for min/max on x64 machines is only available in SSE versions > 3 and AVX.
The "@requires" annotation in the JTREG test has been modified to use the whitelisted flags instead.
Deleting the entry for this JTREG test in test/hotspot/jtreg/ProblemList.txt.