Fix Jtreg failures on NEON with -XX:MaxVectorSize=8
[Byte|Short|Int|Long|Float|Double]128VectorTests.java failed on NEON
platform with -XX:MaxVectorSize=8.

The root cause lies in that 128-bit vector MulReduction operations would
pass the check in match_rule_supported_vector(), i.e. returning true if
lengh_in_bytes is 16, which should be invalid since MaxVectorSize=8.

In this patch, we improve the check logic a bit, i.e. failing fast if
the conditions are not satisfied and falling through to check
vector_size_supported() eventually.

Similar update is done to the check for Op_MulAddVS2VI.