fix: retrofit @throws/@exception partially
**WARNING**: This commit changes the JDK API Documentation.
All changes are of the same kind. One such change can be seen in
java.io.ObjectInputStream.readUTF. That method now documents
previously undocumented EOFException.

The reason why EOFException was not documented previously is
that the old code incorrectly started documentation search
for IOException and its subclasses (for subclasses see 4947455)
from ObjectInputStream.readUTF itself, rather than from the
methods it overrides. Because ObjectInputStream.readUTF documents
IOException, the search succeeded immediately, never finding
EOFException which is documented in java.io.DataInput.readUTF.

This commit changes the code so that the search skips
ObjectInputStream.readUTF when looking for exception
documentation for exceptions listed in the `throws`
clause and their subexceptions.