fix: continue retrofitting @throws/@exception
**WARNING**: This commit changes the JDK API Documentation.

All changes are of the same kind: some methods document more exceptions
without any of the related doc comments having been changed.

For example, along with DateTimeException, ChronoLocalDate.plus(long,
TemporalUnit) now documents UnsupportedTemporalTypeException.

The reason why UnsupportedTemporalTypeException was not documented
previously (although it should've been because of JDK-4947455) is that
ThrowsTaglet.inherit was incorrectly obtaining
javax.lang.model.element.Element corresponding to the exception in
the `@throws` tag. ThrowsTaglet.inherit was looking for a ThrowsTree
node in a wrong DocCommentTree node; what was done like this:

    var ch = utils.getCommentHelper(input.element);
    target = ch.getException(tag);

should've been done like this:

    target = utils.getCommentHelper(input.docTreeInfo.element()).getException(tag);

The new code obtains the target correctly.

The new behavior sometimes results in doubling of exception
documentation. This happens when a subexception is inherited both
explicitly and implicitly when a superexception is inherited.
For example, java.nio.channels.SocketChannel#bind now documents
ClosedChannelException twice.

The first time because:

    @throws  ClosedChannelException              {@inheritDoc}

And the second time because:

    @throws  IOException                         {@inheritDoc}

Such doubling is an intermediate artifact of this commit and will
disappear in subsequent commits.