Replace _unlimited_memory with calls to os::Linux
_unlimited_memory was a constant in cgroupV1Subsystem_linux which is
initialised to a very large number and used as a ceiling sanity check
when reading a number of memory-related cgroup limits. This was not
sufficient to rule out all possible bad values from cgroups and so a
lower ceiling, set to the host's physical RAM, was needed (8292083)

Eliminate _unlimited_memory which is superfluous and use the host
physical memory instead.

For memory_and_swap_limit_in_bytes we need a higher limit than
physical RAM, so extend os::Linux to report on the host's configured
swap value and combine the two.