4668290: unclear spec for Polygon.bounds field
New documentation replaces `This value can be null.`:
```
When created, {@link #invalidate() invalidated}, or {@link #reset() reset},
this becomes {@code null}. To get out of the null state,
{@link #getBounds()} called with {@link #npoints} being greater than
{@code 0} will transfer into a non-null {@link Rectangle}.
```

In javax.swing.text.html.Map, I've replaced setting the field to null with the equivalent #invalidate(). This may mean we can make the field private in the future.