Replace any remaining uses of TreeInfo.isThisQualifier().
Bug 8193904 describes how "A.this.foo" is not detected as an early reference to
the uninitialized final variable "foo". There is also a flip-side to this bug,
which is that "A.this.foo = ..." is not accepted as a valid assignment to the
uninitialized final variable "foo". This commit fixes that part of the bug.

Also do some minor refactoring for clarity.