8298632: [TESTBUG] Add IR checks in jtreg vectorization tests
In JDK-8183390, we introduced a new auto-vectorization testing framework
in hotspot jtreg tests. The new framework provides us a much simpler way
to add new test cases of auto-vectorizable loops. But the previous patch
just added a check of the correctness of the result after vectorization.
There is no check about whether the code is vectorized or not.

This patch adds IR checks to verify C2's vectorization ability on test
cases inside this framework. With this patch, each test method annotated
with `@Test` is verified in two ways. First, it's invoked twice and the
return results from the interpreter and C2 compiled code are compared.
Second, the count of expected vector IR is checked by the IR framework
if the test method has IR rule annotation.

Ideally, we should check IR rules on all platforms. But in practice, the
vectorization ability can be quite different on different platforms, or
different generations of CPUs of one platform. So in this patch, we only
add vectorizable checks for AArch64. Checks for other platforms (such as
x86) can still be added later with more CPU feature conditions. We also
add some negative rules (or in-vectorizable rules) with `@IR(failOn=...`
on cases that should not be vectorized on any platform.

A few more test cases are added within this patch as well.

We tested the new IR rules on below kinds of CPUs.
- AArch64 w/ 512-bit SVE
- AArch64 w/ 128-bit SVE
- AArch64 w/o SVE (NEON only)
- x86