Remove deprecated constructors of URL in SpashScreen
URL constructors have been deprecated in Java 20, and are still used in SplashScreen. This change replaces the constructors of URL to be a URI constructor. Splish splash.
I couldn't verify that finding an image in the jar works in either Java 19 or in a build of this change. `java -splash:image.png -jar path/to/my.jar` failed to show an image on the two versions I tested.