8300227: [macos_aarch64] assert(cpu_has("hw.optional.arm.FEAT_AES")) failed after JDK-8297092
This failure occurs on macOS before xnu-8019 version, mainly because the
sysctlbyname string names were not introduced before xnu-8019. Take the
source code of xnu-7195 [1] as an example.

In this patch, we remove the assertions for the sake of backward
compatibility.

Test:
In my local test environment, one Macmini with xnu-8020.140.41 and
another Macmini with xnu-7195.141.6, fastdebug build can pass now and
the CPU features can be detected as expected.

```
$ ./jdk/bin/java -XX:+PrintFlagsFinal --version | grep SHA
     bool UseSHA                = true       {product} {default}
     bool UseSHA1Intrinsics     = true    {diagnostic} {default}
     bool UseSHA256Intrinsics   = true    {diagnostic} {default}
     bool UseSHA3Intrinsics     = true    {diagnostic} {default}
     bool UseSHA512Intrinsics   = true    {diagnostic} {default}
```

[1] https://github.com/apple-oss-distributions/xnu/blob/rel/xnu-7195/bsd/kern/kern_mib.c#L855