8313592: RISC-V: Link libatomic statically
Currently, RISC-V differs from other platforms in that it requires the linkage to libatomic.so to support sub-word atomic operations. However, because it is linked dynamically, it will depend on the installation of libatomic.so on the system where the Java application will run, which no other platform require.

Instead of dynamically linking, we can statically link so that there is no dependency at run-time as it's already statically linked into libjvm.so.