! Do not "implement" static, interface methods
This commit obsoletes 8186693.

A static method of an interface can neither be inherited nor hidden:
JLS 8.4.8, 8.4.8.2, 9.1.3.

The only affected test is fixed, but the results of `make docs`
slightly differ. This is because before this fix, @see tags were
"inherited" from "hidden" static methods of interfaces.

Here's the list of affected files in the JDK documentation:

  java.base/java/time
    LocalDate.html
    LocalDateTime.html
    ZonedDateTime.html
  java.base/java/time/chrono
    HijrahDate.html
    JapaneseDate.html
    MinguoDate.html
    ThaiBuddhistDate.html

The "inherited" method in all of those files is:

    public static LocalDate from(TemporalAccessor temporal)

While propagating @see information across static methods might be
useful and/or desirable behaviour, it's linked to an incorrect code,
which had to be fixed. If that behavior needs to be re-introduced,
we should investigate it separately.

FWIW, the symptoms of documentation inheritance, such as appearance of
"Description copied from" and inheriting @see information, do NOT
appear even for a validly hidden method: a static method of a class.
So more reasons to fix it for invalidly hidden methods.

Archaeologically relevant bug fixes: 8139101, 8186703.