Remove WorkArounds.overrides(ExecutableElement)
Introduced to solve 8174839, the WorkArounds.overrides method
complicates reasoning and increases the gap between jdk.javadoc and
javax.lang.model. Both qualities are undesirable and could be avoided
if 8174839 is approached differently.

Below are quick analysis and conclusion/suggestion. The analysis is
based on the stacktrace attached to 8174839, the repo snapshot
immediately preceding the commit for 8174839 and the commit itself.

Disclaimer: I haven't verified analysis using the debugger.

Analysis
========

8174839 was caused by an attempt to use a null link generated for the
type mirror corresponding to an element that owns a method
overridden by the given method. Null link could only be generated if
that type mirror was itself null. The type mirror was null because
WorkArounds.overriddenType which was in charge of finding that type
mirror could not find the type whose method overrode the method in
question. Such a method could not be found because
WorkArounds.overriddenType factored in return types,
which were incompatible.

On the other hand, the overridden method itself -- not its owner -- was
successfully found because it was searched by Elements.overrides which
does NOT factor in return types.

The link was generated because an overridden method was found,
but the link was null because the overridden type wasn't found.

So there's a discrepancy/miscoordination.

Conclusion/Suggestion
=====================

8174839 removed discrepancy by changing both search mechanisms to factor
in return types. The result was two nonstandard mechanisms in place of
one standard (jx.l.m.Elements.overrides).

An alternative and probably better solution is to remove the discrepancy
the other way around: by replacing those mechanisms with
Elements.override even though it currently does not factor in return
types.

Incompatible return types for overridden methods is only a problem if
running javadoc on non-compilable source. But javadoc should only be
run on compilable source. If it's not, it shouldn't crash, but it also
shouldn't try to compensate for bad source: garbage in, garbage out.
If Elements.overrides believe the method is overridden, fine, let's
generate a link to the overridden method.

The initial problem was NullPointerException, not problematic source.