Fix the GET_CONTAINER_INFO macro
Previously it would not take in a log_fmt argument and would, therefore,
format OS_CONTAINER_ERROR according to the log format string passed in,
potentially being JULONG_FORMAT. Now with the format passed in
explicitly it's possible to use a different format for the log line
if OS_CONTAINER_ERROR is being returned ('%d'). For the actual log line
in the non-error case no observable differences would be seen.

NOTE: GET_CONTAINER_INFO_LINE would not need the change as it's not
logging OS_CONTAINER_ERROR with a user-supplied format.