Simplify Utils.getFirstVisibleSuperClass
* The removed `typeUtils.isSameType(type, superType)` condition was
historically guarding from passing interfaces or reaching
java.lang.Object. Both cases are handled differently now. Because
the condition had no effect, it was removed. However, when removed,
a couple of issues surfaced and were captured as 8302545 and 8302542.

Until either 8302545 has been resolved or --ignore-source-errors has
been removed, I slightly amended containingModule and containingPackage
methods to account for possible RootPackageSymbol being passed.

Amending Utils.containingPackage was NOT necessary, but was done for
robustness.

Not sure how relevant it is to run javadoc with --ignore-source-errors
though. Again: people should not run javadoc on source that does not
compile. But if they run, javadoc should fail-fast and not try to
recover and proceed no matter what.

* Fixed a typo in the IgnoreSourceErrors test. (FWIW, there used to be
"ignore-errors" options in AOT and build.)