8292914: Drop the counter from lambda class names
The class generated for lambda proxies is now defined as a hidden class. This means that the counter, which was used to ensure a unique class name and avoid clashes, is now redundant. In addition to performing redundant work, this also impacts build reproducibility for native image generators which might already have a strategy to cope with hidden classes but cannot cope with indeterminate definition order for lambda proxy classes.

Additionally, the lambda proxy class dumper is enhanced to prepend a raw timestamp before the filename.