! Simplify Utils.getFirstVisibleSuperClass
Not sure why it was checking for qualified names; perhaps, to avoid
loops, which seem to be guarding against broken source. Loops are
more robustly detected by maintaining a set of already seen
TypeElements.

Anyway, tests pass.

The good news is that some classes now show better type hierarchy.
For example:

    Class ByteVector

    java.lang.Object
      jdk.incubator.vector.Vector<Byte>
                           ^^^^^^^^^^^^
        jdk.incubator.vector.ByteVector

    public abstract class ByteVector
    extends Vector<Byte>
            ^^^^^^^^^^^^

This is similar to what interfaces already had:

    Class StringBuilder

    java.lang.Object
      java.lang.StringBuilder

    All Implemented Interfaces:
    Serializable, Appendable, CharSequence, Comparable<StringBuilder>
                                            ^^^^^^^^^^^^^^^^^^^^^^^^^